+++ index.d.ts
@@ -15,6 +15,7 @@ export interface LineChartProps {
   withOuterLines?: boolean;
   fromZero?: boolean;
   yAxisLabel?: string;
+  yAxisSuffix?: string;
   xAxisLabel?: string;
   chartConfig: ChartConfig;
   decorator?: Function;
@@ -47,7 +48,9 @@ export interface BarChartProps {
   width: number;
   height: number;
   fromZero?: boolean;
+  withInnerLines?: boolean;
   yAxisLabel: string;
+  yAxisSuffix: string;
   chartConfig: ChartConfig;
   style?: object;
   horizontalLabelRotation?: number;
@@ -120,6 +123,8 @@ export interface ChartConfig {
    * Defines the second color opacity in the linear gradient of a chart's background
    */
   backgroundGradientToOpacity?: number;
+  fillShadowGradient?: string;
+  fillShadowGradientOpacity?: number;
   /**
    * Defines the base color function that is used to calculate colors of labels and sectors used in a chart
    */